requirements.txt
streamlit>=1.35.0
PyPDF2>=3.0.0
python-dotenv>=1.0.0
langchain>=0.2.0
langchain-groq>=0.1.0
langchain-community>=0.2.0
faiss-cpu>=1.7.4
sentence-transformers>=2.2.2
